backend/pyproject.toml
[project]
name = "toolchain-backend"
version = "0.1.0"
description = "AI Tool Discovery Platform - Multi-Agent Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "prometheus-client>=0.23.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "redis>=7.1.0",
    "sentry-sdk[fastapi]>=2.49.0",
    "slowapi>=0.1.9",
    "structlog>=25.0.0",
    "tavily-python>=0.5.0",
    "tenacity>=9.0.0",
    "uvicorn>=0.34.0",
    "zhipuai>=2.1.5.20250825",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
